Refactored Repositories.
