+++ GraduationTracker/GraduationTracker/GraduationTracker.csproj
@@ -43,11 +43,16 @@
     <Compile Include="Course.cs" />
     <Compile Include="Diploma.cs" />
     <Compile Include="GraduationTracker.cs" />
+    <Compile Include="Repositories\DiplomaRepository.cs" />
+    <Compile Include="Repositories\RequirementRepository.cs" />
+    <Compile Include="Repositories\StudentRepository.cs" />
     <Compile Include="STANDING.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Repository.cs" />
     <Compile Include="Requirement.cs" />
     <Compile Include="Student.cs" />
+  </ItemGroup>
+  <ItemGroup>
+    <Folder Include="Models\" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
