+++ GraduationTracker/GraduationTracker/Repositories/DiplomaRepository.cs
@@ -1,0 +1,26 @@
+﻿using System.Linq;
+
+namespace GraduationTracker.Repositories
+{
+    public class DiplomaRepository
+    {
+        public static Diploma GetDiploma(int id)
+        {
+            var diplomas = GetDiplomas();
+            return diplomas.Where(d => d.Id == id).FirstOrDefault();
+        }
+
+        private static Diploma[] GetDiplomas()
+        {
+            return new[]
+            {
+                new Diploma
+                {
+                    Id = 1,
+                    Credits = 4,
+                    Requirements = new int[]{100,102,103,104}
+                }
+            };
+        }
+    }
+}
