+++ GraduationTracker/GraduationTracker/Repositories/StudentRepository.cs
@@ -1,0 +1,64 @@
+﻿using System.Linq;
+
+namespace GraduationTracker.Repositories
+{
+    public class StudentRepository
+    {
+        public static Student GetStudent(int id)
+        {
+            var students = GetStudents();
+            return students.Where(s => s.Id == id).FirstOrDefault();
+        }
+
+        private static Student[] GetStudents()
+        {
+            return new[]
+            {
+               new Student
+               {
+                   Id = 1,
+                   Courses = new Course[]
+                   {
+                        new Course{Id = 1, Name = "Math", Mark=95 },
+                        new Course{Id = 2, Name = "Science", Mark=95 },
+                        new Course{Id = 3, Name = "Literature", Mark=95 },
+                        new Course{Id = 4, Name = "Physichal Education", Mark=95 }
+                   }
+               },
+               new Student
+               {
+                   Id = 2,
+                   Courses = new Course[]
+                   {
+                        new Course{Id = 1, Name = "Math", Mark=80 },
+                        new Course{Id = 2, Name = "Science", Mark=80 },
+                        new Course{Id = 3, Name = "Literature", Mark=80 },
+                        new Course{Id = 4, Name = "Physichal Education", Mark=80 }
+                   }
+               },
+               new Student
+               {
+                   Id = 3,
+                   Courses = new Course[]
+                   {
+                       new Course{Id = 1, Name = "Math", Mark=50 },
+                       new Course{Id = 2, Name = "Science", Mark=50 },
+                       new Course{Id = 3, Name = "Literature", Mark=50 },
+                       new Course{Id = 4, Name = "Physichal Education", Mark=50 }
+                   }
+               },
+               new Student
+               {
+                   Id = 4,
+                   Courses = new Course[]
+                   {
+                       new Course{Id = 1, Name = "Math", Mark=40 },
+                       new Course{Id = 2, Name = "Science", Mark=40 },
+                       new Course{Id = 3, Name = "Literature", Mark=40 },
+                       new Course{Id = 4, Name = "Physichal Education", Mark=40 }
+                   }
+               }
+            };
+        }
+    }
+}
